Refactoring duplicated code
