+++ src/Server/UsersHandler.ts
@@ -1,20 +1,14 @@
 import { IncomingMessage, ServerResponse } from "http";
-import { stringify } from "querystring";
 import { UsersDBAccess } from "../User/UsersDBAccess";
-import { Handler, HTTP_CODES, HTTP_METHODS } from "./Model";
-import { Response } from "./Response";
+import { BaseRequestHandler } from "./BaseRequestHandler";
+import { HTTP_METHODS } from "./Model";
 import { Utils } from "./Utils";
 
-export class UserHandler implements Handler {
-  private req: IncomingMessage;
-  private res: ServerResponse;
-  private response: Response;
+export class UserHandler extends BaseRequestHandler {
   private usersDBAccess: UsersDBAccess = new UsersDBAccess();
 
   constructor(req: IncomingMessage, res: ServerResponse) {
-    this.req = req;
-    this.res = res;
-    this.response = new Response();
+    super(req, res);
   }
 
   public async handleRequest(): Promise<void> {
@@ -36,12 +30,5 @@ export class UserHandler implements Handler {
         params[name] = value;
       });
     }
-  }
-
-  private async handleNotFound(): Promise<void> {
-    const data = {
-      message: "Not found",
-    };
-    this.response.createResponse(this.res, HTTP_CODES.NOT_FOUND, data);
   }
 }
